Refactor: controller: move FSA timer-related code to its own source files
